Add requirements.txt for python3.12

+++ utils.py
@@ -150,7 +150,7 @@ class dataset_class(Dataset):
         y = self.labels[ind]  # (num_labels,) array
 
         data = torch.tensor(x, dtype=torch.float32)
-        label = torch.tensor(y, dtype=torch.float32)
+        label = torch.tensor(y)
 
         return data, label, ind
